requirements.txt
alembic==1.6.5
asgiref==3.5.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
Django==3.2.13
Faker==13.7.0
filelock==3.0.12
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-Moment==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.0
idna==3.3
importlib-metadata==3.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
Pillow==9.1.1
psycopg2==2.9.1
pygame==2.0.1
PyJWT==2.1.0
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
PySimpleGUI==4.55.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2022.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.28.1
sip==4.19.8
six==1.16.0
SQLAlchemy==1.4.20
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.10
utmp==20.10.0
virtualenv==20.4.0
Werkzeug==2.0.1
WTForms==2.3.3
XlsxWriter==3.0.2
zipp==3.4.0
